requirements.txt
Django==1.10.3
argparse==1.2.1
django-cors-headers==1.3.1
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
wsgiref==0.1.2
